requirements.txt
xarray
netCDF4
cftime
numpy
pandas
PyYAML
bottleneck
pytest
# optional
dask[complete]
tqdm
matplotlib
seaborn
cartopy
regionmask